Version: 1.3! - added the ability shield; - added a game icon in the form of a bird; - split the clock and shield icons into separate circle and arrow images; - if the bird collides with the left side of the camera or the ceiling, it dies; - lowered the cost of abilities to zero; ------------ Версия: 1.3! - добавил способность щита; - добавил иконку игры в виде птицы; - разбил иконки часов и щита на изображения отдельных круга и стрелок; - при столкновении с левой частью камеры или потолком птица умирает; - опустил стоимость способностей до нуля;

+++ Assets/Scripts/PipeSpawner.cs
@@ -16,11 +16,11 @@ public class PipeSpawner : MonoBehaviour
         while (true)
         {
             yield return new WaitForSeconds(2);
-            SpawnPipe(3, 7);
+            SpawnPipe(3.0f, 7.0f);
         }
     }
 
-    private void SpawnPipe(int rangeStart, int rangeEnd)
+    private void SpawnPipe(float rangeStart, float rangeEnd)
     {
         float rand = Random.Range(rangeStart, rangeEnd);
         GameObject newPipe = Instantiate(Pipe, new Vector3(4, rand, 0), Quaternion.identity);
